+++ plantdescriptionengine/src/main/java/eu/arrowhead/core/plantdescriptionengine/services/pde_mgmt/dto/PlantDescription.java
@@ -18,7 +18,7 @@ import static se.arkalix.dto.DtoEncoding.JSON;
 public interface PlantDescription {
     String plantDescription();
     Optional<Boolean> active();
-    Optional<List<Float>> include();
+    Optional<List<Integer>> include();
     List<PdeSystem> systems();
     List<Connection> connections();
 }
