+++ plantdescriptionengine/src/main/java/eu/arrowhead/core/plantdescriptionengine/services/pde_monitor/dto/PdeAlarm.java
@@ -17,7 +17,7 @@ import static se.arkalix.dto.DtoEncoding.JSON;
 @DtoToString
 public interface PdeAlarm {
 
-    float id();
+    int id();
     String systemName();
     boolean acknowledged();
     String severity();
